--- examples/contrib/pytorch-lightning/seq2seq/__init__.py
@@ -1,5 +1,0 @@
-import os
-import sys
-
-
-sys.path.insert(1, os.path.dirname(os.path.realpath(__file__)))
