+++ examples/seq2seq/test_seq2seq_examples_multi_gpu.py
@@ -3,14 +3,7 @@
 import os
 import sys
 
-from transformers.testing_utils import (
-    TestCasePlus,
-    execute_subprocess_async,
-    get_gpu_count,
-    require_torch_gpu,
-    require_torch_multi_gpu,
-    slow,
-)
+from transformers.testing_utils import TestCasePlus, execute_subprocess_async, get_gpu_count, require_torch_gpu, slow
 
 from .utils import load_json
