+++ research_projects/seq2seq/convert_pl_checkpoint_to_hf.py
