+++ research_projects/seq2seq/test_bash_script.py
